+++ tests/builder/test_architect.py
@@ -4,12 +4,13 @@ from spacekit.builder.architect import BuilderEnsemble, Builder
 from spacekit.skopes.hst.svm.train import load_ensemble_data
 
 
+@mark.hst
 @mark.svm
 @mark.builder
 @mark.architect
-def test_ensemble_builder_with_data(skope, labeled_dataset, svm_train_npz):
+def test_svm_ensemble_builder_with_data(skope, labeled_dataset, svm_train_npz):
     check_skope(skope, "svm")
-    tv_idx, XTR, YTR, XTS, YTS, _, _ = load_ensemble_data(
+    _, XTR, YTR, XTS, YTS, _, _ = load_ensemble_data(
         labeled_dataset,
         svm_train_npz,
         img_size=128,
@@ -41,33 +42,51 @@ def test_ensemble_builder_with_data(skope, labeled_dataset, svm_train_npz):
     assert ens.model is not None
 
 
+@mark.hst
 @mark.svm
 @mark.builder
 @mark.architect
-def test_ensemble_builder_without_data(skope):
+def test_svm_ensemble_builder_without_data(skope):
     check_skope(skope, "svm")
     ens = BuilderEnsemble()
     assert ens.blueprint == "ensemble"
     assert ens.steps_per_epoch > 0
-    ens.load_saved_model(arch="ensemble")
+    ens.load_saved_model(arch="svm_align")
     assert ens.model is not None
     assert len(ens.model.layers) == 28
-    assert ens.model_path == "models/ensemble/ensembleSVM"
+    assert ens.model_path == "models/svm_align/ensembleSVM"
     ens.find_tx_file()
-    assert ens.tx_file == "models/ensemble/tx_data.json"
+    assert ens.tx_file == "models/svm_align/tx_data.json"
 
 
+@mark.hst
 @mark.cal
 @mark.builder
 @mark.architect
-def test_cal_builder_without_data(skope):
-    check_skope(skope, "cal")
+def test_hst_cal_builder_without_data(skope):
+    check_skope(skope, "hstcal")
     builder = Builder(name="mem_clf")
-    builder.load_saved_model("calmodels")
+    builder.load_saved_model("hst_cal")
     assert builder.model is not None
-    assert builder.blueprint == "mem_clf"
+    assert builder.blueprint == "hst_mem_clf"
     builder.get_blueprint(builder.blueprint)
     assert len(builder.model.layers) == 8
-    assert builder.model_path == "models/calmodels/mem_clf"
+    assert builder.model_path == "models/hst_cal/mem_clf"
     builder.find_tx_file()
-    assert builder.tx_file == "models/calmodels/tx_data.json"
+    assert builder.tx_file == "models/hst_cal/tx_data.json"
+
+
+@mark.jwst
+@mark.builder
+@mark.architect
+def test_jwst_cal_builder_without_data(skope):
+    check_skope(skope, "jwstcal")
+    builder = Builder(name="img3_reg")
+    builder.load_saved_model("jwst_cal")
+    assert builder.model is not None
+    assert builder.blueprint == "jwst_img3_reg"
+    builder.get_blueprint(builder.blueprint)
+    assert len(builder.model.layers) == 8
+    assert builder.model_path == "models/jwst_cal/img3_reg"
+    builder.find_tx_file()
+    assert builder.tx_file == "models/jwst_cal/tx_data.json"